+++ packages/coding-agent/src/extensibility/plugins/index.ts
@@ -4,6 +4,7 @@ export * from "@veyyon/kernel/loader/plugins/git-url";
 export * from "@veyyon/kernel/loader/plugins/installed-registry";
 export * from "@veyyon/kernel/loader/plugins/parser";
 export type * from "@veyyon/kernel/loader/plugins/types";
+export type * from "@veyyon/plugin";
 export * from "./doctor";
 export * from "./loader";
 export * from "./manager";