+++ packages/coding-agent/src/extensibility/plugins/loader.ts
@@ -9,12 +9,8 @@ import * as fs from "node:fs";
 import * as path from "node:path";
 import { type ManifestHolder, manifestFromPackageJson } from "@veyyon/kernel/loader/manifest-key";
 import { normalizePluginRuntimeConfig } from "@veyyon/kernel/loader/plugins/runtime-config";
-import type {
-	InstalledPlugin,
-	PluginManifest,
-	PluginRuntimeConfig,
-	ProjectPluginOverrides,
-} from "@veyyon/kernel/loader/plugins/types";
+import type { InstalledPlugin, PluginRuntimeConfig, ProjectPluginOverrides } from "@veyyon/kernel/loader/plugins/types";
+import type { PluginManifest } from "@veyyon/plugin";
 import { errorMessage, getPluginsDir, getPluginsLockfile, isEnoent, logger } from "@veyyon/utils";
 import { getConfigDirPaths } from "../../config";
 import { registerPluginCacheInvalidator, resolveActiveProjectRegistryPath } from "../../discovery/helpers";